#1067 Fix Double -> BigDecimal calling createInventoryItemDetail service
git-svn-id: svn://svn.opentaps.org/opentaps_all/branches/ofbiz-0904-1.4.0-preview1@13033 d3523486-f5fe-0310-a41a-a15a4e76f3c7

+++ hot-deploy/opentaps-common/src/common/org/opentaps/common/domain/inventory/OrderInventoryService.java
@@ -397,7 +397,7 @@ public class OrderInventoryService extends Service implements OrderInventoryServ
                         createDetailMap.put("inventoryItemId", inventoryItem.getInventoryItemId());
                         createDetailMap.put("orderId", orderId);
                         createDetailMap.put("orderItemSeqId", orderItemSeqId);
-                        createDetailMap.put("availableToPromiseDiff", deductAmount.negate().doubleValue());
+                        createDetailMap.put("availableToPromiseDiff", deductAmount.negate());
                         getInfrastructure().getDispatcher().runSync("createInventoryItemDetail", createDetailMap);
 
                         // create OrderItemShipGrpInvRes record
@@ -577,7 +577,7 @@ public class OrderInventoryService extends Service implements OrderInventoryServ
                         context.put("orderId", orderId);
                         context.put("orderItemSeqId", orderItemSeqId);
                         context.put("shipGroupSeqId", shipGroupSeqId);
-                        context.put("availableToPromiseDiff", toReserve.negate().doubleValue());
+                        context.put("availableToPromiseDiff", toReserve.negate());
                         runSync("createInventoryItemDetail", context);
 
                         // calculate the promiseDatetime
@@ -627,7 +627,7 @@ public class OrderInventoryService extends Service implements OrderInventoryServ
                     context.put("orderId", orderId);
                     context.put("orderItemSeqId", orderItemSeqId);
                     context.put("shipGroupSeqId", shipGroupSeqId);
-                    context.put("availableToPromiseDiff", quantityNotReserved.negate().doubleValue());
+                    context.put("availableToPromiseDiff", quantityNotReserved.negate());
                     runSync("createInventoryItemDetail", context);
 
                     // calculate the promiseDatetime
